requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.3.2
drf-spectacular==0.27.0
typing-extensions==4.9.0
